+++ src/main/java/com/backend/allreva/auth/exception/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.auth.handler;
+package com.backend.allreva.auth.exception;
 
 import java.io.IOException;
 
@@ -6,38 +6,51 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.backend.allreva.common.dto.Response;
+import com.backend.allreva.common.exception.CustomException;
+import com.backend.allreva.common.exception.code.ErrorCode;
 import com.backend.allreva.common.exception.code.GlobalErrorCode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Component
+@RequiredArgsConstructor
 public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
     private final ObjectMapper objectMapper;
 
-    public CustomAuthenticationEntryPoint() {
-        this.objectMapper = new ObjectMapper();
-    }
-
     /**
      * 인증 실패 시 호출되는 메서드
      * 
-     * JWT 토큰 인증 실패 시 UNAUTHORIZED 상태로 응답합니다.
+     * JWT 토큰 인증 실패 에러가 발생합니다.
      */
     @Override
     public void commence(HttpServletRequest request, HttpServletResponse response,
             AuthenticationException authException) throws IOException, ServletException {
-        log.info(authException.getMessage());
+        Exception jwtException = (Exception) request.getAttribute("jakarta.servlet.error.exception");
+
+        ErrorCode errorCode;
+        String message;
+
+        if (jwtException != null && jwtException instanceof CustomException) {
+            CustomException customException = (CustomException) jwtException;
+            errorCode = customException.getErrorCode();
+            message = errorCode.message();
+            log.info("JWT 인증 예외 발생: {}", message);
+        } else {
+            errorCode = GlobalErrorCode.UNAUTHORIZED_ERROR.getErrorCode();
+            message = errorCode.message();
+            log.info("기본 인증 예외 발생: {}", message);
+        }
 
         Response<GlobalErrorCode> errorResponse = Response.onFailure(
-                GlobalErrorCode.UNAUTHORIZED_ERROR.getErrorCode().code(),
-                GlobalErrorCode.UNAUTHORIZED_ERROR.getErrorCode().message()
+                errorCode.code(), message
         );
 
         response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
