+++ src/main/java/com/backend/allreva/auth/exception/code/JwtErrorCode.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.auth.exception;
+package com.backend.allreva.auth.exception.code;
 
 import org.springframework.http.HttpStatus;
 
@@ -10,9 +10,9 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum JwtErrorCode implements ErrorCodeInterface {
 
-    INVALID_JWT_TOKEN(HttpStatus.BAD_REQUEST.value(), "INVALID_JWT_TOKEN", "유효하지 않은 토큰입니다."),
-    INVALID_JWT_SIGNATURE(HttpStatus.BAD_REQUEST.value(), "INVALID_JWT_SIGNATURE", "유효하지 않은 서명입니다."),
-    EXPIRED_JWT_TOKEN(HttpStatus.BAD_REQUEST.value(), "EXPIRED_JWT_TOKEN", "토큰이 만료되었습니다.")
+    INVALID_JWT_TOKEN(HttpStatus.UNAUTHORIZED.value(), "INVALID_JWT_TOKEN", "유효하지 않은 토큰입니다."),
+    INVALID_JWT_SIGNATURE(HttpStatus.UNAUTHORIZED.value(), "INVALID_JWT_SIGNATURE", "유효하지 않은 서명입니다."),
+    EXPIRED_JWT_TOKEN(HttpStatus.UNAUTHORIZED.value(), "EXPIRED_JWT_TOKEN", "토큰이 만료되었습니다.")
     ;
 
     private final Integer status;
